Set up frontend for integration with WebGL
 * Currently doing really dumb stuff like re-compiling the shader every tick
 * Browser locks up completely for some reason
   * Switched to Asm.JS to try to get a better idea of what's going on
 * Fixed formats of the data so that WebGL is happy
   * Internal buffer is now a `Vec<f32>` instead of a `Vec<f64>`
 * Fixed my copy-and-paste of the OpenGL code

+++ volumetric/engine/src/noise_middleware.rs
@@ -39,7 +39,7 @@ fn drive_noise<C: CellState + BufColumn>(
                 let val = noise.get([x as f64 * zoom, y as f64 * zoom, (z as f64) * speed]);
 
                 // set the cell's state equal to that value
-                column[z] = val;
+                column[z] = val as f32;
             }
         }
     }
